refactor(userProfile): change UserProfile component structure
